requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.34.0
Pillow==9.5.0
nltk==3.8.1
numpy<2.0
streamlit==1.21.0
pdf2image==1.16.0
python-dotenv==1.0.0
